+++ src/lib.rs
@@ -1,7 +1,9 @@
 mod common;
 mod completion;
+mod config;
 mod document;
 mod logger;
 mod parser;
 mod server;
+mod utils;
 pub use crate::server::PathServer;