+++ src/server.rs
@@ -1,31 +1,35 @@
 use std::collections::{HashMap, HashSet};
-use std::sync::Arc;
+use std::path::PathBuf;
 
 use tokio::sync::Mutex;
 use tokio::sync::RwLock;
-use tower_lsp::jsonrpc::Result;
+use tower_lsp::jsonrpc;
 use tower_lsp::lsp_types;
 
 use crate::common::*;
 use crate::completion;
+use crate::config;
 use crate::document::Document;
 use crate::logger::{self, *};
 use crate::parser;
+use crate::utils::url_to_path;
+
+const VERSION: &str = env!("CARGO_PKG_VERSION");
 
 #[derive(Debug)]
 pub struct PathServer {
-    // client: tower_lsp::Client,
-    workspace_roots: RwLock<HashSet<lsp_types::Url>>,
-    documents: Arc<Mutex<HashMap<String, Document>>>, // url -> document
+    client: tower_lsp::Client,
+    workspace_roots: RwLock<HashSet<PathBuf>>,
+    documents: Mutex<HashMap<PathBuf, Document>>, // url -> document
 }
 
 impl PathServer {
     pub fn new(client: tower_lsp::Client) -> Self {
         logger::init(&client);
         Self {
-            // client,
+            client,
             workspace_roots: RwLock::new(HashSet::new()),
-            documents: Arc::new(Mutex::new(HashMap::new())),
+            documents: Mutex::new(HashMap::new()),
         }
     }
 }
@@ -35,11 +39,30 @@ impl tower_lsp::LanguageServer for PathServer {
     async fn initialize(
         &self,
         params: lsp_types::InitializeParams,
-    ) -> Result<lsp_types::InitializeResult> {
-        if let Some(url) = params.root_uri {
-            log(format!("Adding workspace root: {}", url)).await;
+    ) -> jsonrpc::Result<lsp_types::InitializeResult> {
+        // for backward compatibility
+        if let Some(uri) = params.root_uri {
+            let Ok(root) = url_to_path(&uri) else {
+                info(format!("Failed to convert root URI to file path: {}", uri)).await;
+                return Err(jsonrpc::Error::invalid_params("Invalid root URI"));
+            };
             let mut roots = self.workspace_roots.write().await;
-            roots.insert(url.clone());
+            roots.insert(root);
+        }
+        if let Some(folders) = params.workspace_folders {
+            let mut roots = self.workspace_roots.write().await;
+            for folder in folders {
+                log(format!("Adding workspace root: {}", folder.uri)).await;
+                let Ok(root) = url_to_path(&folder.uri) else {
+                    info(format!(
+                        "Failed to convert URI to file path: {}",
+                        folder.uri
+                    ))
+                    .await;
+                    continue;
+                };
+                roots.insert(root);
+            }
         }
         Ok(lsp_types::InitializeResult {
             capabilities: lsp_types::ServerCapabilities {
@@ -71,6 +94,11 @@ impl tower_lsp::LanguageServer for PathServer {
 
     async fn initialized(&self, _: lsp_types::InitializedParams) {
         log("Path Server initialized".to_string()).await;
+        log(format!("Path Server version: {}", VERSION)).await;
+    }
+
+    async fn did_change_configuration(&self, _: lsp_types::DidChangeConfigurationParams) {
+        // TODO: implement it
     }
 
     async fn did_change_workspace_folders(
@@ -80,27 +108,46 @@ impl tower_lsp::LanguageServer for PathServer {
         for folder in params.event.added {
             log(format!("Adding workspace folder: {}", folder.uri)).await;
             let mut roots = self.workspace_roots.write().await;
-            roots.insert(folder.uri.clone());
+            let Ok(root) = url_to_path(&folder.uri) else {
+                info(format!(
+                    "Failed to convert URI to file path: {}",
+                    folder.uri
+                ))
+                .await;
+                continue;
+            };
+            roots.insert(root);
         }
         for folder in params.event.removed {
             log(format!("Removing workspace folder: {}", folder.uri)).await;
             let mut roots = self.workspace_roots.write().await;
-            roots.remove(&folder.uri);
+            let Ok(root) = url_to_path(&folder.uri) else {
+                continue;
+            };
+            roots.remove(&root);
         }
     }
 
     async fn did_open(&self, params: lsp_types::DidOpenTextDocumentParams) {
-        self.documents.lock().await.insert(
-            params.text_document.uri.to_string(),
-            Document::new(params.text_document.text),
-        );
+        let mut documents = self.documents.lock().await;
+        let Ok(path) = url_to_path(&params.text_document.uri) else {
+            info(format!(
+                "Failed to convert URI to file path: {}",
+                params.text_document.uri
+            ))
+            .await;
+            return;
+        };
+        documents.insert(path, Document::new(params.text_document.text));
     }
 
     async fn did_change(&self, params: lsp_types::DidChangeTextDocumentParams) {
-        let uri = params.text_document.uri.to_string();
+        let Ok(path) = url_to_path(&params.text_document.uri) else {
+            return;
+        };
         let mut docs = self.documents.lock().await;
         let doc = docs
-            .entry(uri)
+            .entry(path)
             .or_insert_with(|| Document::new(String::new()));
         // apply each change in order
         for change in params.content_changes.into_iter() {
@@ -119,23 +166,35 @@ impl tower_lsp::LanguageServer for PathServer {
     }
 
     async fn did_close(&self, params: lsp_types::DidCloseTextDocumentParams) {
-        self.documents
-            .lock()
-            .await
-            .remove(&params.text_document.uri.to_string());
+        let Ok(path) = url_to_path(&params.text_document.uri) else {
+            return;
+        };
+        self.documents.lock().await.remove(&path);
     }
 
     async fn completion(
         &self,
         params: lsp_types::CompletionParams,
-    ) -> Result<Option<lsp_types::CompletionResponse>> {
+    ) -> jsonrpc::Result<Option<lsp_types::CompletionResponse>> {
         // get the line prefix
         let line_number = params.text_document_position.position.line as usize;
         let character = params.text_document_position.position.character as usize;
-        let url = &params.text_document_position.text_document.uri.to_string();
+        let Ok(path) = url_to_path(&params.text_document_position.text_document.uri) else {
+            info(format!(
+                "Failed to convert URI to file path: {}",
+                params.text_document_position.text_document.uri
+            ))
+            .await;
+            return Ok(None);
+        };
         let documents = self.documents.lock().await;
-        let Some(doc) = documents.get(url) else {
-            return Err(PathServerError::Unknown(format!("Document not found: {}", url)).into());
+        let Some(doc) = documents.get(&path) else {
+            info(format!("Document not found: {}", path.display())).await;
+            return Err(PathServerError::Unknown(format!(
+                "Document not found: {}",
+                path.display()
+            ))
+            .into());
         };
         let line_prefix = doc.get_line(line_number, Some(character))?;
 
@@ -144,14 +203,24 @@ impl tower_lsp::LanguageServer for PathServer {
         debug(format!("Completing for prefix: '{}'", raw_path)).await;
 
         // completion
-        let file_path = params.text_document_position.text_document.uri;
+        let completion_config = config::get(&self.client).await.completion;
+        let Ok(file_path) = url_to_path(&params.text_document_position.text_document.uri) else {
+            info(format!(
+                "Failed to convert URI to file path: {}",
+                params.text_document_position.text_document.uri
+            ))
+            .await;
+            return Ok(None);
+        };
         let workspace_roots = self.workspace_roots.read().await;
-        let completions = completion::complete(&raw_path, &workspace_roots, &file_path).await?;
+        let completions =
+            completion::complete(&raw_path, &workspace_roots, &file_path, &completion_config)
+                .await?;
 
         return Ok(Some(lsp_types::CompletionResponse::Array(completions)));
     }
 
-    async fn shutdown(&self) -> Result<()> {
+    async fn shutdown(&self) -> jsonrpc::Result<()> {
         Ok(())
     }
 }